Split component
